+++ src/message/index.js
@@ -8,60 +8,101 @@
  */
 
 import './index.less'
-import React, { Component } from 'react'
-import { render } from 'react-dom'
-import ReactCSSTransitionGroup from 'react-addons-css-transition-group'
-import invariant from 'invariant'
+import React, { PropTypes, Component } from 'react'
+import ReactDOM from 'react-dom'
 import classnames from 'classnames'
+import ToggleNode from '../_shared/ToggleNode'
 import Icon from '../Icon'
+import Button from '../Button'
 
 class Message extends Component {
 
-  constructor() {
+  constructor(props) {
     super()
-    this.state = {}
+    this.state = {
+      open: props.open
+    }
+    this.prepareClose(props)
+  }
+
+  componentDidMount() {
+    this.toggleNode = new ToggleNode(ReactDOM.findDOMNode(this), 'bfd-message--open')
+    this.toggleNode.onClose = () => {
+      this.props.onClose && this.props.onClose()
+    }
+    this.toggleNode.open()
+  }
+
+  componentWillReceiveProps(nextProps) {
+    'open' in nextProps && this.setState({open: nextProps.open})
+    this.prepareClose(nextProps)
+  }
+
+  componentDidUpdate() {
+    this.toggleNode[this.state.open ? 'open' : 'close']()
+  }
+
+  prepareClose(props) {
+    const { duration } = props
+    duration && setTimeout(::this.handleClose, duration * 1000)
+  }
+
+  handleClose() {
+    this.setState({open: false})
   }
 
   render() {
-    const { show, type, message } = this.state
+    const { type, message, duration } = this.props
     return (
-      <ReactCSSTransitionGroup transitionName="bfd-message--in" transitionEnterTimeout={200} transitionLeaveTimeout={150}>
-        {show && (
-          <div className={classnames('bfd-message', {[`bfd-message--${type}`]: type})}>
-            <Icon
-              className="bfd-message__symbol"
-              type={type === 'success' ? 'check': 'warning'}
-            />
-            {message}
-          </div>
+      <div className={classnames('bfd-message', {[`bfd-message--${type}`]: type})}>
+        <Icon
+          className="bfd-message__symbol"
+          type={type === 'success' ? 'check': 'warning'}
+        />
+        {message}
+        {duration === 0 && (
+          <Button
+            className="bfd-message__remove"
+            transparent
+            icon="remove"
+            onClick={::this.handleClose}
+          />
         )}
-      </ReactCSSTransitionGroup>
+      </div>
     )
   }
 }
 
-let instance
+Message.propTypes = {
+  type: PropTypes.oneOf(['success', 'danger']).isRequired,
+  message: PropTypes.node.isRequired,
+  duration: PropTypes.number,
+  onClose: PropTypes.func,
+  open: PropTypes.bool
+}
 
-const showMessage = (type, message, duration) => {
-  if (process.env.NODE_ENV !== 'production') {
-    invariant(typeof message === 'string' || (message && React.isValidElement(message)), '`message` should be `string` or `ReactElement`, check the first param of message.' + type)
+let render = props => {
+  const container = document.createElement('div')
+  document.body.appendChild(container)
+
+  let isOpen = false
+  const messageQueue = []
+  const handleClose = () => {
+    isOpen = false
+    const head = messageQueue[0] && messageQueue.shift()
+    head && render(head)
   }
-  if (!instance) {
-    const container = document.createElement('div')
-    document.body.appendChild(container)
-    instance = render(<Message />, container)
+
+  render = nextProps => {
+    props = Object.assign({}, props, nextProps)
+    if (isOpen && props.open) {
+      return messageQueue.push(props)
+    }
+    isOpen = props.open
+    ReactDOM.render(<Message {...props} onClose={handleClose} />, container)
+    props.open || handleClose()
   }
-  if (!instance.state.show) {
-    instance.setState({
-      message,
-      type,
-      duration,
-      show: true
-    })
-    setTimeout(() => {
-      instance.setState({show: false})
-    }, duration * 1000)
-  }
+  render()
 }
 
 const message = {
@@ -70,31 +111,41 @@ const message = {
    * @public
    * @name message.success
    * @param  {string | element} message message 内容，支持 React 元素
-   * @param  {number} [duration] 持续时间，单位毫秒
+   * @param  {number} [duration] 持续时间，单位秒，为0时手动关闭
    * @description 成功信息，默认 2 秒后自动关闭
    */
   success(message, duration = 2) {
-    showMessage('success', message, duration)
+    render({
+      message,
+      duration,
+      type: 'success',
+      open: true
+    })
   },
 
   /**
    * @public
    * @name message.danger
    * @param  {string | element} message message 内容，支持 React 元素
-   * @param  {number} [duration] 持续时间，单位毫秒
+   * @param  {number} [duration] 持续时间，单位秒，为0时手动关闭
    * @description 失败信息，默认 3 秒后自动关闭
    */
   danger(message, duration = 3) {
-    showMessage('danger', message, duration)
+    render({
+      message,
+      duration,
+      type: 'danger',
+      open: true
+    })
   },
 
   /**
    * @public
    * @name message.close
-   * @description 关闭 message 信息
+   * @description 关闭当前 message
    */
   close() {
-    instance && instance.setState({show: false})
+    render({open: false})
   }
 }